main/link_shortener/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pyjwt
bcrypt
requests
